refactor: 중복된 타입 분리
