+++ frontend/src/@types/index.ts
@@ -1,0 +1,9 @@
+interface Schedule {
+  id: number;
+  title: string;
+  startDateTime: string;
+  endDateTime: string;
+  memo: string;
+}
+
+export { Schedule };
